Event Code Printing - Logging the different event codes - Adding in cache and requests for the VIN API call - Spaces cleanup

+++ Extensions/StringExtensions.cs
@@ -5,46 +5,46 @@ namespace CSharpInterviewMessageProcessor.Extensions;
 public static class StringExtensions
 {
     /// <summary>
-    /// Convert string version of long value to long.
+    ///     Convert string version of double value to double.
     /// </summary>
     /// <param name="value">string number</param>
-    /// <returns>long value or long.MinValue</returns>
-    public static long ToLong(this string value)
+    /// <returns>double value or double.MinValue</returns>
+    public static double ToDouble(this string value)
     {
-        return long.TryParse(value, out var output) 
-            ? output 
-            : long.MinValue;
+        return double.TryParse(value, out var output)
+            ? output
+            : double.MinValue;
     }
-    
-    
+
+
     /// <summary>
-    /// Convert string version of date to DateTimeOffset.
+    ///     Convert string version of date to DateTimeOffset.
     /// </summary>
     /// <param name="value">string number</param>
     /// <returns>int value or int.MinValue</returns>
     public static DateTimeOffset ToDateTimeOffset(this string value)
     {
-        return DateTime.TryParse(value, out var output) 
-            ? output 
+        return DateTime.TryParse(value, out var output)
+            ? output
             : DateTimeOffset.MinValue;
     }
-    
-    
+
+
     /// <summary>
-    /// Convert string version of int value to int.
+    ///     Convert string version of int value to int.
     /// </summary>
     /// <param name="value">string number</param>
     /// <returns>int value or int.MinValue</returns>
     public static int ToInt(this string? value)
     {
-        return int.TryParse(value, out var output) 
-            ? output 
+        return int.TryParse(value, out var output)
+            ? output
             : int.MinValue;
     }
-    
-    
+
+
     /// <summary>
-    /// Fluent expression for checking null of whitespace.
+    ///     Fluent expression for checking null of whitespace.
     /// </summary>
     public static bool IsNotNullOrWhiteSpace(
         this string? input)
